Reorder and deprecate
