--- src/DataBlobs/entities/AbstractDataStore.jl
@@ -1,7 +1,0 @@
-
-"""
-    $(TYPEDEF)
-Superclass of all key-value datastores.
-"""
-abstract type AbstractDataStore{T}
-end
