+++ src/DataBlobs/services/AbstractDataEntries.jl
@@ -1,22 +1,40 @@
-##
-
-getHash(entry::AbstractDataEntry) = hex2bytes(entry.hash)
-
-
-##
+##==============================================================================
+## AbstractDataEntry - compare
+##==============================================================================
 
 import Base: ==
 
-function ==(a::GeneralDataEntry, b::GeneralDataEntry)
-    return a.label == b.label &&
-        a.storeKey == b.storeKey &&
-        a.mimeType == b.mimeType &&
-        Dates.value(a.createdTimestamp - b.createdTimestamp) < 1000 &&
-        Dates.value(a.lastUpdatedTimestamp - b.lastUpdatedTimestamp) < 1000 #1 second
+@generated function ==(x::T, y::T) where T <: AbstractDataEntry
+    mapreduce(n -> :(x.$n == y.$n), (a,b)->:($a && $b), fieldnames(x))
 end
 
-@generated function ==(x::MongodbDataEntry, y::MongodbDataEntry)
-    mapreduce(n -> :(x.$n == y.$n), (a,b)->:($a && $b), fieldnames(x))
+##==============================================================================
+## AbstractDataEntry - common
+##==============================================================================
+
+"""
+    $(SIGNATURES)
+Function to generate source string - userId|robotId|sessionId|varLabel
+"""
+buildSourceString(dfg::AbstractDFG, label::Symbol) =
+    "$(dfg.userId)|$(dfg.robotId)|$(dfg.sessionId)|$label"
+
+
+##==============================================================================
+## AbstractDataEntry - CRUD
+##==============================================================================
+
+"""
+    $(SIGNATURES)
+Get big data entry
+"""
+function getDataEntry(var::AbstractDFGVariable, key::Symbol)
+    !hasDataEntry(var, key) && error("Data entry $(key) does not exist in variable")
+    return var.dataDict[key]
+end
+
+function getDataEntry(dfg::AbstractDFG, label::Symbol, key::Symbol)
+    return getDataEntry(getVariable(dfg, label), key)
 end
 
 """
@@ -39,113 +57,6 @@ function addDataEntry!(dfg::AbstractDFG, label::Symbol, bde::AbstractDataEntry)
     return addDataEntry!(getVariable(dfg, label), bde)
 end
 
-
-"""
-$(SIGNATURES)
-Add Big Data Entry to distributed factor graph.
-Should be extended if DFG variable is not returned by reference.
-
-Example
-
-See docs for `getDataEntryElement`.
-
-Related
-
-addData!, getDataEntryElement, fetchData
-"""
-function addData!(dfg::AbstractDFG,
-                  lbl::Symbol,
-                  datastore::Union{FileDataStore, InMemoryDataStore},
-                  descr::Symbol,
-                  mimeType::AbstractString,
-                  data::Vector{UInt8} )
-  #
-  node = isVariable(dfg, lbl) ? getVariable(dfg, lbl) : getFactor(dfg, lbl)
-  # Make a big data entry in the graph - use JSON2 to just write this
-  entry = GeneralDataEntry(dfg, node, descr, mimeType=mimeType)
-  # Set it in the store
-  addDataBlob!(datastore, entry, data)
-  # Add the entry to the graph
-  addDataEntry!(node, entry)
-end
-# const addDataEntry! = addData!
-
-"""
-    $SIGNATURES
-
-Does a data entry (element) exist at `key`.
-"""
-hasDataEntry(var::DFGVariable, key::Symbol) = haskey(var.dataDict, key)
-
-"""
-    $(SIGNATURES)
-Get big data entry
-"""
-function getDataEntry(var::AbstractDFGVariable, key::Symbol)
-    !hasDataEntry(var, key) && error("Data entry $(key) does not exist in variable")
-    return var.dataDict[key]
-end
-
-function getDataEntry(dfg::AbstractDFG, label::Symbol, key::Symbol)
-    return getDataEntry(getVariable(dfg, label), key)
-end
-
-"""
-$SIGNATURES
-Get both the entry and raw data element from datastore returning as a tuple.
-
-Notes:
-- This is the counterpart to `addDataEntry!`.
-- Data is identified by the node in the DFG object `dfglabel::Symbol` as well as `datalabel::Symbol`.
-- The data should have been stored along with a `entry.mimeType::String` which describes the format of the data.
-- ImageMagick.jl is useful for storing images in png or jpg compressed format.
-
-Example
-
-```julia
-# some dfg object
-fg = initfg()
-addVariable!(fg, :x0, IIF.ContinuousScalar) # using IncrementalInference
-# FileDataStore (as example)
-datastore = FileDataStore(joinLogPath(fg,"datastore"))
-
-# now some data comes in
-mydata = Dict(:soundBite => randn(Float32, 10000), :meta => "something about lazy foxes and fences.")
-
-# store the data
-addDataEntry!( fg, :x0, datastore, :SOUND_DATA, "application/json", Vector{UInt8}(JSON2.write( mydata ))  )
-
-# get/fetch the data
-entry, rawData = fetchData(fg, :x0, datastore, :SOUND_DATA)
-
-# unpack data to original format (this must be done by the user)
-@show entry.mimeType # "applicatio/json"
-userFormat = JSON2.read(IOBuffer(rawData))
-```
-
-Related
-
-addDataEntry!, addData!, fetchData, fetchDataEntryElement
-"""
-function getDataEntryBlob(dfg::AbstractDFG,
-                          dfglabel::Symbol,
-                          datastore::Union{FileDataStore, InMemoryDataStore},
-                          datalabel::Symbol)
-  #
-  vari = getVariable(dfg, dfglabel)
-  if !hasDataEntry(vari, datalabel)
-      # current standards is to fail hard
-    error("missing data entry $datalabel in $dfglabel")
-    # return nothing, nothing
-  end
-  entry = getDataEntry(vari, datalabel)
-  element = getDataBlob(datastore, entry)
-  return entry, element
-end
-const fetchDataEntryElement = getDataEntryBlob
-const fetchData = getDataEntryBlob
-
-#
 
 """
     $(SIGNATURES)
@@ -186,6 +97,18 @@ function deleteDataEntry!(var::AbstractDFGVariable, entry::AbstractDataEntry)
     return deleteDataEntry!(var, entry.label)
 end
 
+##==============================================================================
+## AbstractDataEntry - Helper functions, Lists, etc
+##==============================================================================
+
+"""
+    $SIGNATURES
+
+Does a data entry (element) exist at `key`.
+"""
+hasDataEntry(var::DFGVariable, key::Symbol) = haskey(var.dataDict, key)
+
+
 """
     $(SIGNATURES)
 Get big data entries, Vector{AbstractDataEntry}
@@ -199,7 +122,6 @@ function getDataEntries(dfg::AbstractDFG, label::Symbol)
     #or should we return the iterator, Base.ValueIterator{Dict{Symbol,AbstractDataEntry}}?
     getDataEntries(getVariable(dfg, label))
 end
-
 
 """
     $(SIGNATURES)
